requirements.txt
streamlit>=1.44
pillow>=9.5
opencv-python >= 4.12.0
numpy >=2.2.1
av >=14.4.0
requests >= 2.32.2
pandas >= 2.2.3
fpdf >= 1.7.2
tensorflow >= 2.20.0
matplotlib >= 3.10.0
scikit-learn >= 1.6.1